% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_ci}
\alias{combine_ci}
\title{Combine bounds for presentation.}
\usage{
combine_ci(df, label)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{label}{name of the new column}
}
\value{

}
\description{
Combine bounds for presentation.
}
