% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_digits}
\alias{format_digits}
\title{Format digits to be the same length}
\usage{
format_digits(x, n)
}
\arguments{
\item{x}{vector to which apply}

\item{n}{the number of digits to follow the decimal}
}
\value{

}
\description{
Format digits to be the same length
}
