% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tidy.skim_df}
\alias{tidy.skim_df}
\title{Tidy a numeric skimr table.}
\usage{
\method{tidy}{skim_df}(
  x,
  missing = FALSE,
  complete = TRUE,
  n = TRUE,
  mean = TRUE,
  sd = TRUE,
  p0 = FALSE,
  p25 = FALSE,
  p50 = FALSE,
  p75 = FALSE,
  p100 = FALSE,
  hist = FALSE
)
}
\arguments{
\item{x}{an object of class skim_df}

\item{missing}{return the missing column, logical}

\item{complete}{return the complete column, logical}

\item{n}{return the n column, logical}

\item{mean}{return the mean column, logical}

\item{sd}{return the sd column, logical}

\item{p0}{return the p0 column, logical}

\item{p25}{return the p25 column, logical}

\item{p50}{return the p50 column, logical}

\item{p75}{return the p75 column, logical}

\item{p100}{return the p100 column, logical}

\item{hist}{return the hist column, logical}
}
\value{

}
\description{
Tidy a numeric skimr table.
}
